[FE] style: GroupCardList 분리
- group card list 컴포넌트 분리
